refactor/document util.hh
